Refactoring AgentProcess: - AgentProcess history, memory, get_input has been deprecated - history moved to it's own module for later integration into Langchain memory class - AgentProcess refactored for updated chat model. Now requires task and chain. - AgentProcess.start() now expects an inputs dict that is passed through to the chain - AgentProcess now records agent.alias for THINK messages - lots of updated and deprecated tests - moved fixtures to conftest for reuse elsewhere
