+++ ix/agents/tests/test_process.py
@@ -1,12 +1,11 @@
 from typing import List
-from unittest.mock import MagicMock
 
 import pytest
 import json
 
 from ix.agents.exceptions import AgentQuestion
 from ix.agents.process import DEFAULT_MEMORY, AgentProcess
-from ix.agents.tests.mock_llm import MockChatOpenAI
+from ix.conftest import USER_INPUT
 from ix.memory.redis import RedisVectorMemory
 from ix.task_log.models import TaskLogMessage
 from ix.task_log.tests.fake import (
@@ -22,38 +21,6 @@ from ix.task_log.tests.fake import (
 
 
 NEXT_COMMAND = {"user_input": "NEXT COMMAND"}
-
-
-@pytest.fixture
-def task():
-    return fake_task()
-
-
-@pytest.fixture
-def task_log_msg(task):
-    return fake_task_log_msg(task)
-
-
-@pytest.fixture
-def mock_openai(mocker):
-    # create a mock instance of the class
-    mock_llm = MockChatOpenAI()
-
-    # mock the class to return the instance we're creating here
-    mock_class = MagicMock(return_value=mock_llm)
-
-    # mock_import returns the mock class
-    mock_import = mocker.patch("ix.agents.llm.import_llm_class")
-    mock_import.return_value = mock_import.return_value = mock_class
-
-    # return the mock instance
-    yield mock_llm
-
-
-@pytest.fixture()
-def command_output(mocker):
-    """mocks write_output to capture output from `echo` command"""
-    yield mocker.patch("ix.agents.tests.echo_command.write_output")
 
 
 @pytest.fixture()
@@ -78,7 +45,6 @@ class MockTicker:
         # use toggle to stop autonomous mode
         if not self.remaining:
             fake_autonomous_toggle(task=self.task, enabled=0)
-            self.agent_process.update_message_history()
 
         # simulate return code, will be False when autonomous mode
         # is disabled and agent returns with AUTH_REQUEST
@@ -96,202 +62,8 @@ class MessageTeardown:
 @pytest.mark.django_db
 class TestAgentProcessInit:
     def test_init_with_defaults(self, task):
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         assert agent_process.task_id == task.id
-        assert agent_process.memory_class == DEFAULT_MEMORY
-
-
-@pytest.mark.django_db
-class TestAgentProcessHistory(MessageTeardown):
-    MSG_1 = {"type": "COMMAND", "message": "THIS IS A TEST 1"}
-    MSG_2 = {"type": "COMMAND", "message": "THIS IS A TEST 2"}
-    MSG_3 = {"type": "COMMAND", "message": "THIS IS A TEST 3"}
-    MSG_4 = {"type": "COMMAND", "message": "THIS IS A TEST 4"}
-
-    @pytest.mark.parametrize("content_type", AgentProcess.EXCLUDED_MSG_TYPES)
-    def test_query_message_excludes_msgs(self, content_type, task):
-        """Test that non-relevant messages are excluded from history"""
-        msg = fake_task_log_msg_type(content_type, task=task)
-        agent_process = AgentProcess(task_id=task.id)
-        agent_process.update_message_history()
-
-        for history_msg in agent_process.history:
-            assert history_msg != msg.as_message()
-
-    @pytest.mark.parametrize(
-        "content_type", ["FEEDBACK", "COMMAND", "EXECUTED", "EXECUTE_ERROR"]
-    )
-    def test_query_message_includes_types(self, content_type, task):
-        """Test that relevant messages are included in history"""
-        msg = fake_task_log_msg_type(content_type, task=task)
-        agent_process = AgentProcess(task_id=task.id)
-        agent_process.update_message_history()
-        assert len(agent_process.history) >= 1
-        assert agent_process.history[-1] == msg.as_message()
-
-    def test_query_message_history_without_since(self, task):
-        """Test querying all messages"""
-        msg1 = fake_task_log_msg(task=task)
-        msg2 = fake_task_log_msg(task=task)
-        agent_process = AgentProcess(task_id=task.id)
-        messages = agent_process.query_message_history()
-        assert messages.count() == 2
-        assert messages.filter(id=msg1.id).exists()
-        assert messages.filter(id=msg2.id).exists()
-
-    def test_query_message_history_with_since(self, task):
-        """Test querying new messages"""
-        msg1 = fake_task_log_msg(task=task)
-        msg2 = fake_task_log_msg(task=task)
-        agent_process = AgentProcess(task_id=task.id)
-        messages = agent_process.query_message_history(since=msg1.created_at)
-        assert messages.count() == 1
-        assert not messages.filter(id=msg1.id).exists()
-        assert messages.filter(id=msg2.id).exists()
-
-    def test_update_message_history_no_messages(self, task):
-        """initializing history before there are messages"""
-        agent_process = AgentProcess(task_id=task.id)
-        assert len(agent_process.history) == 0
-
-        # still no messages
-        agent_process.update_message_history()
-        assert len(agent_process.history) == 0
-
-        # first update
-        msg1 = fake_command_reply(task=task)
-        msg2 = fake_command_reply(task=task)
-        agent_process.update_message_history()
-        assert len(agent_process.history) == 2
-        assert agent_process.history[0] == msg1.as_message()
-        assert agent_process.history[1] == msg2.as_message()
-
-    def test_update_message_history_with_messages(self, task):
-        """initialize history when there are messages"""
-        assert not TaskLogMessage.objects.all().exists()
-        msg1 = fake_task_log_msg(task=task, content=self.MSG_1).as_message()
-        msg2 = fake_task_log_msg(task=task, content=self.MSG_2).as_message()
-        agent_process = AgentProcess(task_id=task.id)
-        agent_process.update_message_history()
-        assert len(agent_process.history) == 2
-        assert agent_process.history[0] == msg1
-        assert agent_process.history[1] == msg2
-
-        # updating with no new messages
-        agent_process.update_message_history()
-        assert len(agent_process.history) == 2
-        assert agent_process.history[0] == msg1
-        assert agent_process.history[1] == msg2
-
-        # add more messages and update
-        msg3 = fake_task_log_msg(task=task, content=self.MSG_3).as_message()
-        msg4 = fake_task_log_msg(task=task, content=self.MSG_4).as_message()
-        agent_process.update_message_history()
-        assert len(agent_process.history) == 4
-        assert agent_process.history[0] == msg1
-        assert agent_process.history[1] == msg2
-        assert agent_process.history[2] == msg3
-        assert agent_process.history[3] == msg4
-
-    def test_continuous_toggle(self, task, mock_openai, command_output):
-        """Can enable/disable autonomous mode"""
-        assert not TaskLogMessage.objects.all().exists()
-        fake_task_log_msg(task=task, content=self.MSG_1)
-        fake_task_log_msg(task=task, content=self.MSG_2)
-        agent_process = AgentProcess(task_id=task.id)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-        # toggle autonomous to enabled
-        fake_autonomous_toggle(enabled=1)
-        agent_process.update_message_history()
-        assert agent_process.autonomous
-
-        # toggle autonomous to disabled
-        fake_autonomous_toggle(enabled=0)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-        # toggle autonomous to enabled
-        fake_autonomous_toggle(enabled=1)
-        agent_process.update_message_history()
-        assert agent_process.autonomous
-
-        # toggle autonomous to disabled
-        fake_autonomous_toggle(enabled=0)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-    def test_autonomous_toggle_latest(self, task):
-        """only latest toggle is used"""
-        assert not TaskLogMessage.objects.all().exists()
-        fake_task_log_msg(task=task, content=self.MSG_1)
-        fake_task_log_msg(task=task, content=self.MSG_2)
-        agent_process = AgentProcess(task_id=task.id)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-        # verify it uses latest with many back and forth
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=0)
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=0)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-        # verify it uses latest
-        fake_autonomous_toggle(enabled=0)
-        fake_autonomous_toggle(enabled=1)
-        agent_process.update_message_history()
-        assert agent_process.autonomous
-
-        # verify it uses latest
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=0)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-        # verify it uses latest with many back and forth
-        fake_autonomous_toggle(enabled=0)
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=0)
-        fake_autonomous_toggle(enabled=1)
-        agent_process.update_message_history()
-        assert agent_process.autonomous
-
-    def test_autonomous_toggle_interspersed(self, task, mock_openai, command_output):
-        """Doesn't have to be the latest of all messages"""
-        assert not TaskLogMessage.objects.all().exists()
-        fake_task_log_msg(task=task, content=self.MSG_1)
-        fake_task_log_msg(task=task, content=self.MSG_2)
-        agent_process = AgentProcess(task_id=task.id)
-        agent_process.update_message_history()
-        assert not agent_process.autonomous
-
-        fake_task_log_msg(task=task, content=self.MSG_1)
-        fake_autonomous_toggle(enabled=1)
-        fake_autonomous_toggle(enabled=0)
-        fake_autonomous_toggle(enabled=1)
-        fake_task_log_msg(task=task, content=self.MSG_1)
-        agent_process.update_message_history()
-        assert agent_process.autonomous
-
-
-@pytest.mark.django_db
-class TestAgentProcessMemory:
-    def test_init_with_custom_memory_class(self, task):
-        custom_memory_class = "ix.memory.tests.mock_vector_memory.MockMemory"
-        assert (
-            custom_memory_class != DEFAULT_MEMORY
-        ), "test shouldn't use the default memory as the custom option"
-        agent_process = AgentProcess(task_id=task.id, memory_class=custom_memory_class)
-        assert agent_process.memory_class == custom_memory_class
-
-    def test_initialize_memory_with_default_memory_class(self, task):
-        agent_process = AgentProcess(task_id=task.id)
-        assert isinstance(agent_process.memory, RedisVectorMemory)
 
 
 @pytest.mark.django_db
@@ -302,9 +74,9 @@ class TestAgentProcessStart:
         mock_reply.delete()
         query = TaskLogMessage.objects.filter(task=task)
         assert query.count() == 0
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         mock_openai.return_value = msg_to_response(mock_reply)
-        return_value = agent_process.start()
+        return_value = agent_process.start(USER_INPUT)
         assert return_value is True
         assert query.count() == 2
 
@@ -312,22 +84,6 @@ class TestAgentProcessStart:
         thought_msg = query[1]
         assert think_msg.content["type"] == "THINK"
         assert thought_msg.content["type"] == "THOUGHT"
-
-    def test_restart_task_with_auth_for_none(self, task, mock_openai):
-        """Restarting task when task is waiting for AUTHORIZE does nothing"""
-        mock_reply = fake_command_reply(task=task)
-        feedback_request = fake_feedback_request(task=task, message_id=mock_reply.id)
-        query = TaskLogMessage.objects.filter(
-            task=task, created_at__gt=feedback_request.created_at
-        )
-
-        agent_process = AgentProcess(task_id=task.id)
-        mock_openai.return_value = msg_to_response(mock_reply)
-
-        # start process
-        return_value = agent_process.start()
-        assert return_value is False
-        assert query.count() == 0
 
     def test_restart_task_with_feedback(self, task, mock_openai, mock_embeddings):
         """
@@ -342,13 +98,11 @@ class TestAgentProcessStart:
         )
         assert query.count() == 0
 
-        agent_process = AgentProcess(
-            task_id=task.id,
-        )
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         mock_openai.return_value = msg_to_response(mock_reply)
 
         # start process
-        return_value = agent_process.start()
+        return_value = agent_process.start(USER_INPUT)
         assert return_value is True
         assert query.count() == 2
         think_msg = query[0]
@@ -364,7 +118,7 @@ class TestAgentProcessStart:
         yet reached the maximum number of ticks or is in autonomous mode.
         """
         fake_autonomous_toggle(task=task, enabled=1)
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         agent_process.tick = MockTicker(agent_process, task, return_value=False)
         return_value = agent_process.start(n=3)
         assert return_value is False
@@ -398,7 +152,7 @@ class TestAgentProcessTicks:
         mock_reply.delete()
         query = TaskLogMessage.objects.filter(task=task)
         assert query.count() == 0
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         mock_openai.return_value = msg_to_response(mock_reply)
         agent_process.tick(user_input={"user_input": "extra input"}, execute=False)
         assert query.count() == 2
@@ -419,7 +173,7 @@ class TestAgentProcessTicks:
         mock_reply.delete()
         query = TaskLogMessage.objects.filter(task=task)
         assert query.count() == 0
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         mock_openai.return_value = msg_to_response(mock_reply)
         agent_process.tick(user_input={"user_input": "extra input"}, execute=True)
 
@@ -435,7 +189,7 @@ class TestAgentProcessTicks:
         mock_reply.delete()
         query = TaskLogMessage.objects.filter(task=task)
         assert query.count() == 0
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         mock_openai.raise_exception = AgentQuestion("this is a fake question")
         return_value = agent_process.tick(NEXT_COMMAND, execute=True)
 
@@ -455,63 +209,6 @@ class TestAgentProcessTicks:
         assert msg_1.content["type"] == "FEEDBACK_REQUEST"
         assert msg_1.content["question"] == "this is a fake question"
 
-    def test_tick_updates_history(self, task, mock_openai, mock_embeddings):
-        """Tick starts by updating history, each tick should load any new messages"""
-        mock_reply = fake_command_reply(task=task)
-        mock_reply.delete()
-        query = TaskLogMessage.objects.filter(task=task)
-        assert query.count() == 0
-        agent_process = AgentProcess(task_id=task.id)
-        mock_response = msg_to_response(mock_reply)
-        mock_openai.return_value = mock_response
-
-        mock_chat_message_content = (
-            '{"command": {"args": {"output": "this is a test"}, "name": "echo"}, "thoughts": {"criticism": '
-            '"constructive self-criticism", "plan": ["short list of steps", "that conveys", "long-term plan"], '
-            '"reasoning": "reasoning", "speak": "thoughts summary to say to user", "text": "thought"}}'
-        )
-        mock_ai_chat_message = {
-            "content": mock_chat_message_content,
-            "role": "assistant",
-        }
-        mock_user_chat_message = {
-            "content": NEXT_COMMAND,
-            "role": "user",
-        }
-
-        # verify that history is loaded on each tick. Note that history from a tick
-        # isn't loaded until the next tick.
-
-        # first tick starts with no history
-        agent_process.tick(NEXT_COMMAND, execute=False)
-        message_types = list(query.values_list("content__type", flat=True))
-        assert message_types == ["THINK", "THOUGHT"]
-        assert agent_process.history == []
-
-        # tick again adds THINK and COMMAND from first tick
-        agent_process.tick(NEXT_COMMAND, execute=False)
-        message_types = list(query.values_list("content__type", flat=True))
-        assert message_types == [
-            "THINK",
-            "THOUGHT",
-            "THINK",
-            "THOUGHT",
-        ]
-        assert agent_process.history == [
-            mock_user_chat_message,
-            # mock_ai_chat_message,
-        ]
-
-        # tick with new messages
-        agent_process.tick(NEXT_COMMAND, execute=False)
-        assert query.all().count() == 6
-        assert agent_process.history == [
-            mock_user_chat_message,
-            # mock_ai_chat_message,
-            mock_user_chat_message,
-            # mock_ai_chat_message,
-        ]
-
 
 @pytest.mark.django_db
 class TestAgentProcessAIChat:
@@ -524,7 +221,7 @@ class TestAgentProcessAIChat:
         think_msg = fake_think()
 
         mock_openai.return_value = msg_to_response(mock_reply)
-        agent_process = AgentProcess(task_id=task.id)
+        agent_process = AgentProcess(task_id=task.id, chain_id=task.chain.id)
         message = {"user_input": "Test message"}
         agent_process.chat_with_ai(think_msg, message)
         # mock_openai.assert_called()
