+++ ix/schema/mutations/chat.py
@@ -146,7 +146,7 @@ class ChatInputMutation(graphene.Mutation):
 
         # Start agent loop. This does NOT check if the loop is already running
         # the agent_runner task is responsible for blocking duplicate runners
-        start_agent_loop.delay(str(input.task_id), message_id=message.id)
+        start_agent_loop.delay(str(task_id), str(agent.chain.id), inputs=inputs)
 
         return TaskLogMessageResponse(task_log_message=message)
